Partway through fixing general setup instructions
