--- src/simple_uam/tools/setup_lin/__init__.py
@@ -1,7 +1,0 @@
-"""
-SimpleUAM windows node setup scripts.
-"""
-
-from typing import List # noqa
-
-__all__: List[str] = []  # noqa: WPS410 (the only __variable__ we use)
