--- src/simple_uam/tools/setup_lin/__main__.py
@@ -1,15 +1,0 @@
-"""
-Entry-point module, in case you use `python -m util`.
-
-Why does this file exist, and why `__main__`? For more info, read:
-
-- https://www.python.org/dev/peps/pep-0338/
-- https://docs.python.org/3/using/cmdline.html#cmdoption-m
-"""
-
-import sys
-
-from .cli import main
-
-if __name__ == "__main__":
-    sys.exit(main(sys.argv[1:]))
